move multihop fns to its own file and simplify fn signature
