+++ modules/core/03-connection/keeper/verify.go
@@ -13,7 +13,6 @@ import (
 	channeltypes "github.com/cosmos/ibc-go/v8/modules/core/04-channel/types"
 	commitmenttypes "github.com/cosmos/ibc-go/v8/modules/core/23-commitment/types"
 	host "github.com/cosmos/ibc-go/v8/modules/core/24-host"
-	mh "github.com/cosmos/ibc-go/v8/modules/core/33-multihop"
 	ibcerrors "github.com/cosmos/ibc-go/v8/modules/core/errors"
 	"github.com/cosmos/ibc-go/v8/modules/core/exported"
 )
@@ -365,154 +364,6 @@ func (k Keeper) VerifyNextSequenceRecv(
 	}
 
 	return nil
-}
-
-// VerifyMultihopMembership verifies a multi-hop membership proof.
-func (k Keeper) VerifyMultihopMembership(
-	ctx sdk.Context,
-	connection exported.ConnectionI,
-	height exported.Height,
-	proof []byte,
-	connectionHops []string,
-	kvGenerator channeltypes.KeyValueGenFunc,
-) error {
-	var multihopProof channeltypes.MsgMultihopProofs
-	if err := k.cdc.Unmarshal(proof, &multihopProof); err != nil {
-		return err
-	}
-
-	// get the last hop connection on the other side of the multihop channel
-	// the last hop connection is the connection end on the chain before the counterparty multihop chain
-	lastHopConnectionEnd, err := multihopProof.GetLastHopConnectionEnd(k.cdc, connection)
-	if err != nil {
-		return err
-	}
-
-	// generate the key/value for the expected value on the counterparty end that needs to be proven
-	key, value, err := kvGenerator(&multihopProof, lastHopConnectionEnd)
-	if err != nil {
-		return errorsmod.Wrap(err, "failed to generate key and value to prove")
-	}
-
-	// the counterparty of the last hop connection end is the connection end on the other end of the multihop channel
-	prefix := lastHopConnectionEnd.GetCounterparty().GetPrefix()
-	path, err := commitmenttypes.ApplyPrefix(prefix, commitmenttypes.NewMerklePath(key))
-	if err != nil {
-		return err
-	}
-
-	clientID := connection.GetClientID()
-	clientStore := k.clientKeeper.ClientStore(ctx, clientID)
-
-	clientState, found := k.clientKeeper.GetClientState(ctx, clientID)
-	if !found {
-		return errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID)
-	}
-
-	// check client associated with connection on this end of the multihop channel is active
-	if status := clientState.Status(ctx, clientStore, k.cdc); status != exported.Active {
-		return errorsmod.Wrapf(clienttypes.ErrClientNotActive, "client (%s) status is %s", clientID, status)
-	}
-
-	if clientState.GetLatestHeight().LT(height) {
-		return errorsmod.Wrapf(
-			clienttypes.ErrInvalidHeight,
-			"client state height < proof height (%d < %d), please ensure the client has been updated",
-			clientState.GetLatestHeight(), height,
-		)
-	}
-
-	delayPeriod, err := multihopProof.GetMaximumDelayPeriod(k.cdc, connection)
-	if err != nil {
-		return err
-	}
-
-	expectedTimePerBlock := k.GetParams(ctx).MaxExpectedTimePerBlock
-
-	// ensure the delayPeriod passed
-	if err := mh.VerifyDelayPeriodPassed(ctx, clientStore, height, delayPeriod, expectedTimePerBlock); err != nil {
-		return err
-	}
-
-	consensusState, found := k.clientKeeper.GetClientConsensusState(ctx, clientID, height)
-	if !found {
-		return errorsmod.Wrapf(clienttypes.ErrConsensusStateNotFound, "consensus state not found for client ID (%s) at height (%s)", clientID, height)
-	}
-
-	return mh.VerifyMultihopMembership(k.cdc, consensusState, connectionHops, &multihopProof, path, value)
-}
-
-// VerifyMultihopNonMembership verifies a multi-hop non-membership proof.
-func (k Keeper) VerifyMultihopNonMembership(
-	ctx sdk.Context,
-	connection exported.ConnectionI,
-	height exported.Height,
-	proof []byte,
-	connectionHops []string,
-	kvGenerator channeltypes.KeyGenFunc,
-) error {
-	var mProof channeltypes.MsgMultihopProofs
-	if err := k.cdc.Unmarshal(proof, &mProof); err != nil {
-		return err
-	}
-
-	lastHopConnectionEnd, err := mProof.GetLastHopConnectionEnd(k.cdc, connection)
-	if err != nil {
-		return err
-	}
-
-	// generate the key on the counterparty end that needs to be proven
-	key, err := kvGenerator(&mProof, lastHopConnectionEnd)
-	if err != nil {
-		return errorsmod.Wrap(err, "failed to generate key")
-	}
-
-	// the counterparty of the last hop connection end is the connection end on the other end of the multihop channel
-	prefix := lastHopConnectionEnd.GetCounterparty().GetPrefix()
-	path, err := commitmenttypes.ApplyPrefix(prefix, commitmenttypes.NewMerklePath(key))
-	if err != nil {
-		return err
-	}
-
-	clientID := connection.GetClientID()
-	clientStore := k.clientKeeper.ClientStore(ctx, clientID)
-
-	clientState, found := k.clientKeeper.GetClientState(ctx, clientID)
-	if !found {
-		return errorsmod.Wrap(clienttypes.ErrClientNotFound, clientID)
-	}
-
-	// check client associated with connection on this end of the multihop channel is active
-	if status := clientState.Status(ctx, clientStore, k.cdc); status != exported.Active {
-		return errorsmod.Wrapf(clienttypes.ErrClientNotActive, "client (%s) status is %s", clientID, status)
-	}
-
-	if clientState.GetLatestHeight().LT(height) {
-		return errorsmod.Wrapf(
-			clienttypes.ErrInvalidHeight,
-			"client state height < proof height (%d < %d), please ensure the client has been updated",
-			clientState.GetLatestHeight(), height,
-		)
-	}
-
-	delayPeriod, err := mProof.GetMaximumDelayPeriod(k.cdc, connection)
-	if err != nil {
-		return err
-	}
-
-	expectedTimePerBlock := k.GetParams(ctx).MaxExpectedTimePerBlock
-
-	// ensure the delayPeriod passed
-	if err := mh.VerifyDelayPeriodPassed(ctx, clientStore, height, delayPeriod, expectedTimePerBlock); err != nil {
-		return err
-	}
-
-	consensusState, found := k.clientKeeper.GetClientConsensusState(ctx, clientID, height)
-	if !found {
-		return errorsmod.Wrapf(clienttypes.ErrConsensusStateNotFound, "consensus state not found for client ID (%s) at height (%s)", clientID, height)
-	}
-
-	return mh.VerifyMultihopNonMembership(k.cdc, consensusState, connectionHops, &mProof, path)
 }
 
 // getBlockDelay calculates the block delay period from the time delay of the connection
